+++ CrossEngine/src/RenderSystem/RenderGLES/CrossGLES3DescriptorSet.cpp
@@ -71,7 +71,7 @@ namespace CrossEngine {
 		m_uniformBufferOffsets[binding] = 0;
 	}
 
-	void CGLES3DescriptorSet::SetUniformBufferData(uint32_t binding, uint32_t offset, uint32_t size, const void *pBuffer)
+	void CGLES3DescriptorSet::SetUniformBufferData(uint32_t binding, size_t offset, size_t size, const void *pBuffer)
 	{
 		if (m_ptrUniformBuffers.find(binding) != m_ptrUniformBuffers.end()) {
 			m_ptrUniformBuffers[binding]->SetData(offset, size, pBuffer);
